Renames activities queries to match express routes #127
