+++ react-backend/database/queries/activities_today.js
