+++ react-backend/database/queries/activity_add.js
