+++ react-backend/database/queries/activity_delete.js
