+++ react-backend/database/queries/activity_update.js
