+++ react-backend/routes/activities_add.js
@@ -1,12 +1,12 @@
 const express = require('express');
-const insertActivity = require('../database/queries/insertActivity');
+const activityAdd = require('../database/queries/activity_add');
 
 const router = express.Router();
 
 router.post('/', (req, res, next) => {
   const activityToAdd = req.body.name;
-  insertActivity(activityToAdd, req.auth.cb_id)
-    .then(id => res.send({ id, token: req.auth.adminToken }))
+  activityAdd(activityToAdd, req.auth.cb_id)
+    .then(id => res.send({ id }))
     .catch(next);
 });
