+++ react-backend/routes/activities_all.js
@@ -1,16 +1,11 @@
 const express = require('express');
-const getActivities = require('../database/queries/activities');
+const activities = require('../database/queries/activities');
 
 const router = express.Router();
 
 router.get('/', (req, res, next) => {
-  console.log(req.auth);
-
-  getActivities(req.auth.cb_id)
-    .then(activities => {
-      console.log(activities);
-      res.send({ activities, token: req.auth.adminToken });
-    })
+  activities(req.auth.cb_id)
+    .then(activities => res.send({ activities }))
     .catch(next);
 });
