+++ react-backend/routes/activities_delete.js
@@ -1,11 +1,11 @@
 const express = require('express');
-const deleteActivity = require('../database/queries/deleteActivity');
+const activityDelete = require('../database/queries/activity_delete');
 
 const router = express.Router();
 
 router.post('/', (req, res, next) => {
-  deleteActivity(req.body.id, req.auth.cb_id)
-    .then(() => res.send({ success: 'success', token: req.auth.adminToken }))
+  activityDelete(req.body.id, req.auth.cb_id)
+    .then(() => res.send({ success: 'success' }))
     .catch(next);
 });
