+++ react-backend/routes/activities_update.js
@@ -1,11 +1,11 @@
 const express = require('express');
-const updateActivity = require('../database/queries/updateActivity');
+const activityUpdate = require('../database/queries/activity_update');
 
 const router = express.Router();
 
 router.post('/', (req, res, next) => {
   const activity = req.body;
-  updateActivity(
+  activityUpdate(
     activity.id,
     activity.monday,
     activity.tuesday,
@@ -16,7 +16,7 @@ router.post('/', (req, res, next) => {
     activity.sunday,
     req.auth.cb_id
   )
-    .then(() => res.send({ success: 'success', token: req.auth.adminToken }))
+    .then(() => res.send({ success: 'success' }))
     .catch(next);
 });
